# docs/static/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Epistimology API
  version: 1.0.0
  description: >-
    Reporting and argument-audit API: debates, full-text content, full-text search logging, and audit items
    (contraarguments, fallacies, evidences, sources, interpretations).
paths:
...
  /debates:
    post:
      operationId: post-debates
      summary: Create a debate report
      description: Stores a new report as an article with paragraphs, reusing or
      tags:
        - Debates
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DebateInput'
        required: true
      responses:
        '201':
          description: Report stored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebateResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Storage failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /search:
    get:
      operationId: get-search
      summary: Log a search query
      description: Validates the `q` query parameter, truncates it to 200
      tags:
        - Search
      parameters:
        - in: query
          name: q
          required: false
          schema:
            type: string
            description: Full-text search query. Truncated server-side to 200 characters.
          description: Full-text search query. Truncated server-side to 200 characters.
          example: example
      responses:
        '200':
          description: Logged query
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          description: Invalid query
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
servers:
  - url: ''
    description: API server
components:
  schemas:
...
    DebateInput:
      type: object
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 200
          description: Report title, 3-200 characters
        label:
          type: string
          maxLength: 60
          description: Reporting tab label. Reuses the tab when it exists, otherwise creates it.
        description:
          type: string
          minLength: 1
          maxLength: 20000
          description: Markdown report body. Blank-line separated paragraphs become stored paragraphs.
        tabId:
          type: string
          minLength: 1
          description: Existing reporting tab ID. Takes precedence over label.
      required:
        - title
        - description
    DebateResponse:
      type: object
      properties:
        articleId:
          type: string
          description: Created reporting article ID
        tabId:
          type: string
          description: Reporting tab the article was stored under
        paragraphIds:
          type: array
          items:
            type: string
          description: Stored paragraph IDs in document order
      required:
        - articleId
        - tabId
        - paragraphIds
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
      required:
        - error
...
    SearchResponse:
      type: object
      properties:
        query:
          type: string
          description: Query after server-side truncation
        logged:
          type: boolean
          description: Whether the query was logged
      required:
        - query
        - logged
...